Extract `ModelDetailHeader` component
