+++ frontend/src/metabase/models/components/ModelDetailPage/ModelDetailHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ModelDetailHeader";
